more fixing warnings and stuff
Other stuff that is bad about this code, besides the spacing and lack
of brackets I mentioned in my previous commit:
- original author apparently could not tell the difference between '='
and '==', so I am having a hard time guessing what their original
intent was in deciding which one to use...
- uses 'goto' way too much. Normally '-Wunused-label' should not
produce this many warnings, because there should not be that many
labels for 'goto' statements to jump to in the first place

Anyways I think I might have screwed it up even more by trying to fix
all this stuff...

+++ ModuleCmd_List.c
@@ -90,7 +90,7 @@ static	char	_proc_ModuleCmd_List[] = "ModuleCmd_List";
  ** ************************************************************************ **
  ++++*/
 
-int	ModuleCmd_List(Tcl_Interp	*interp, int argc, char *argv[])
+int	ModuleCmd_List(Tcl_Interp *interp, int argc, char *argv[])
 {
     /**
      **  Get the list of loaded modules at first
@@ -157,7 +157,7 @@ int	ModuleCmd_List(Tcl_Interp	*interp, int argc, char *argv[])
 		 **/
 
 		for ((i = 0); (i < count1); i++) {
-			len = (strlen(files[i]) - strlen(list[i]));
+			len = ((int)strlen(files[i]) - (int)strlen(list[i]));
 			tmplist[i] = files[i];
 
 			/**
@@ -167,7 +167,7 @@ int	ModuleCmd_List(Tcl_Interp	*interp, int argc, char *argv[])
 
 			s = (files[i] + len);
 			while (s) {
-				if (s = strchr( s, '/')) {
+				if (s = strchr(s, '/')) {
 					*s = '\0';
 				}
